Trying to make a universal scenario struct (make_x86 doesn't work)

+++ test/nmpu0/textureTriangle/compareImages.h
@@ -1,1 +1,1 @@
-int compareImages(nm32s *input,nm32s *gold,int w, int h);
+double compareImages(nm32s *input,nm32s *gold,int w, int h);
